Fix win64 (size_t vs int in CHECK).
TBR=jarin@chromium.org
BUG=

Review URL: https://codereview.chromium.org/485813003

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@23315 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/compiler/verifier.cc
@@ -307,7 +307,7 @@ void ScheduleVerifier::Run(Schedule* schedule) {
   CHECK_EQ(start, rpo_order->at(0));  // Start should be first.
   for (size_t b = 0; b < rpo_order->size(); b++) {
     BasicBlock* block = rpo_order->at(b);
-    CHECK_EQ(b, block->rpo_number_);
+    CHECK_EQ(static_cast<int>(b), block->rpo_number_);
     BasicBlock* dom = schedule->dominator(block);
     if (b == 0) {
       // All blocks except start should have a dominator.
